% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_gene_exp}
\alias{yeast_gene_exp}
\title{Yeast gene expression}
\format{
A tibble with 199,296 rows and 4 columns:
\describe{
\item{systematic_name}{systematic gene name}
\item{nutrient}{limiting nutrient}
\item{rate}{growth rate}
\item{exp}{gene expression value}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/17959824/}
}
\usage{
yeast_gene_exp
}
\description{
Expression of \emph{S. cerevisiae} genes during chemostat growth under
nutrient limitation.
}
\seealso{
\url{http://varianceexplained.org/files/Brauer2008_DataSet1.tds}
}
\keyword{datasets}
