% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_go_terms}
\alias{yeast_go_terms}
\title{Yeast gene ontology}
\format{
A tibble with 5,537 rows and 4 columns
\describe{
\item{systematic_name}{systematic gene name}
\item{common_name}{common gene name}
\item{go_molecular_function}{GO molecular function}
\item{go_biological_process}{GO biological process}
}
}
\usage{
yeast_go_terms
}
\description{
Gene ontology terms for \emph{S. cerevisiae} genes
}
\seealso{
\url{https://geneontology.org/docs/ontology-documentation/}
}
\keyword{datasets}
