% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{yeast_prot_props}
\alias{yeast_prot_props}
\title{Yeast protein properties}
\format{
A tibble with 6,724 rows and 38 columns
\describe{
\item{systematic_name}{systematic gene name}
\item{mw}{molecular weight}
\item{pi}{isoelectric point}
...
}
}
\source{
http://sgd-archive.yeastgenome.org/curation/calculated_protein_info/protein_properties.tab
}
\usage{
yeast_prot_props
}
\description{
Protein properties for \emph{S. cerevisiae} genes
}
\seealso{
http://sgd-archive.yeastgenome.org/curation/calculated_protein_info/protein_properties.README
}
\keyword{datasets}
