% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yeastgenes-package.R
\docType{package}
\name{yeastgenes-package}
\alias{yeastgenes}
\alias{yeastgenes-package}
\title{yeastgenes: Yeast gene expression and associated protein function data}
\description{
Provides gene expression data from Brauer et al. (PMID 17959824) for *Saccharomyces cerevisiae* grown under nutrient limitation in a chemostat. Also provides data from the Saccharomyces Genome Database.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://rnabioco.github.io/yeastgenes}
}

}
\author{
\strong{Maintainer}: Jay Hesselberth \email{jay.hesselberth@cuanschutz.edu} (\href{https://orcid.org/0000-0002-6299-179X}{ORCID})

}
\keyword{internal}
